refactor: upload modify
