+++ packages/core/src/shader/Shader.ts
@@ -1,26 +1,28 @@
-import { GLCapabilityType } from "../base/Constant";
 import { Engine } from "../Engine";
-import { ShaderFactory } from "../shaderlib/ShaderFactory";
 import { ShaderDataGroup } from "./enums/ShaderDataGroup";
 import { ShaderMacro } from "./ShaderMacro";
 import { ShaderMacroCollection } from "./ShaderMacroCollection";
-import { ShaderProgram } from "./ShaderProgram";
+import { ShaderPass } from "./ShaderPass";
 import { ShaderProperty } from "./ShaderProperty";
 
 /**
- * Shader containing vertex and fragment source.
+ * Shader for rendering.
  */
 export class Shader {
   /** @internal */
   static readonly _compileMacros: ShaderMacroCollection = new ShaderMacroCollection();
+  /** @internal */
+  static readonly _shaderExtension: string[] = [
+    "GL_EXT_shader_texture_lod",
+    "GL_OES_standard_derivatives",
+    "GL_EXT_draw_buffers"
+  ];
 
-  private static _shaderCounter: number = 0;
   private static _shaderMap: Record<string, Shader> = Object.create(null);
   private static _propertyNameMap: Record<string, ShaderProperty> = Object.create(null);
   private static _macroMaskMap: string[][] = [];
   private static _macroCounter: number = 0;
   private static _macroMap: Record<string, ShaderMacro> = Object.create(null);
-  private static _shaderExtension = ["GL_EXT_shader_texture_lod", "GL_OES_standard_derivatives", "GL_EXT_draw_buffers"];
 
   /**
    * Create a shader.
@@ -28,12 +30,21 @@ export class Shader {
    * @param vertexSource - Vertex source code
    * @param fragmentSource - Fragment source code
    */
-  static create(name: string, vertexSource: string, fragmentSource: string): Shader {
+  static create(name: string, vertexSource: string, fragmentSource: string): Shader;
+
+  /**
+   * Create a shader.
+   * @param name - Name of the shader
+   * @param shaderPasses - Shader passe
+   */
+  static create(name: string, shaderPasses: ShaderPass[]): Shader;
+
+  static create(name: string, vertexSourceOrShaderPasses: string | ShaderPass[], fragmentSource?: string): Shader {
     const shaderMap = Shader._shaderMap;
     if (shaderMap[name]) {
       throw `Shader named "${name}" already exists.`;
     }
-    return (shaderMap[name] = new Shader(name, vertexSource, fragmentSource));
+    return (shaderMap[name] = new Shader(name, vertexSourceOrShaderPasses, fragmentSource));
   }
 
   /**
@@ -104,7 +115,10 @@ export class Shader {
     return shaderProperty?._group;
   }
 
-  private static _getNamesByMacros(macros: ShaderMacroCollection, out: string[]): void {
+  /**
+   * @internal
+   */
+  static _getNamesByMacros(macros: ShaderMacroCollection, out: string[]): void {
     const maskMap = Shader._macroMaskMap;
     const mask = macros._mask;
     out.length = 0;
@@ -124,16 +138,18 @@ export class Shader {
   readonly name: string;
 
   /** @internal */
-  _shaderId: number = 0;
+  _shaderPasses: ShaderPass[] = [];
 
-  private _vertexSource: string;
-  private _fragmentSource: string;
-
-  private constructor(name: string, vertexSource: string, fragmentSource: string) {
-    this._shaderId = Shader._shaderCounter++;
+  private constructor(name: string, vertexSourceOrShaderPasses: string | ShaderPass[], fragmentSource?: string) {
     this.name = name;
-    this._vertexSource = vertexSource;
-    this._fragmentSource = fragmentSource;
+
+    if (typeof vertexSourceOrShaderPasses === "string") {
+      this._shaderPasses.push(new ShaderPass(vertexSourceOrShaderPasses, fragmentSource));
+    } else {
+      for (let i = 0, n = vertexSourceOrShaderPasses.length; i < n; i++) {
+        this._shaderPasses.push(vertexSourceOrShaderPasses[i]);
+      }
+    }
   }
 
   /**
@@ -152,64 +168,12 @@ export class Shader {
     for (let i = 0, n = macros.length; i < n; i++) {
       compileMacros.enable(Shader.getMacroByName(macros[i]));
     }
-    return this._getShaderProgram(engine, compileMacros).isValid;
-  }
 
-  /**
-   * @internal
-   */
-  _getShaderProgram(engine: Engine, macroCollection: ShaderMacroCollection): ShaderProgram {
-    const shaderProgramPool = engine._getShaderProgramPool(this);
-    let shaderProgram = shaderProgramPool.get(macroCollection);
-    if (shaderProgram) {
-      return shaderProgram;
+    let isValid = true;
+    const shaderPasses = this._shaderPasses;
+    for (let i = 0, n = shaderPasses.length; i < n; i++) {
+      isValid &&= shaderPasses[i]._getShaderProgram(engine, compileMacros).isValid;
     }
-
-    const isWebGL2: boolean = engine._hardwareRenderer.isWebGL2;
-    const macroNameList = [];
-    Shader._getNamesByMacros(macroCollection, macroNameList);
-    const macroNameStr = ShaderFactory.parseCustomMacros(macroNameList);
-    const versionStr = isWebGL2 ? "#version 300 es" : "#version 100";
-    let precisionStr = `
-    #ifdef GL_FRAGMENT_PRECISION_HIGH
-      precision highp float;
-      precision highp int;
-    #else
-      precision mediump float;
-      precision mediump int;
-    #endif
-    `;
-
-    if (engine._hardwareRenderer.canIUse(GLCapabilityType.shaderTextureLod)) {
-      precisionStr += "#define HAS_TEX_LOD\n";
-    }
-    if (engine._hardwareRenderer.canIUse(GLCapabilityType.standardDerivatives)) {
-      precisionStr += "#define HAS_DERIVATIVES\n";
-    }
-
-    let vertexSource = ShaderFactory.parseIncludes(
-      ` ${versionStr}
-        ${precisionStr}
-        ${macroNameStr}
-        ` + this._vertexSource
-    );
-
-    let fragmentSource = ShaderFactory.parseIncludes(
-      ` ${versionStr}
-        ${isWebGL2 ? "" : ShaderFactory.parseExtension(Shader._shaderExtension)}
-        ${precisionStr}
-        ${macroNameStr}
-      ` + this._fragmentSource
-    );
-
-    if (isWebGL2) {
-      vertexSource = ShaderFactory.convertTo300(vertexSource);
-      fragmentSource = ShaderFactory.convertTo300(fragmentSource, true);
-    }
-
-    shaderProgram = new ShaderProgram(engine, vertexSource, fragmentSource);
-
-    shaderProgramPool.cache(shaderProgram);
-    return shaderProgram;
+    return isValid;
   }
 }
